pyproject.toml
[project]
name = "bio-showcase"
version = "2025.5.17.0"
description = "Soda's Bioinformatics showcase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.2.0",
    "djangorestframework>=3.16.0",
    "primer3-py>=2.1.0",
    "soda-nginx>=2025.5.13.5 ; sys_platform == 'win32'",
    "soda-update-version>=2025.5.17.4",
    "sodatools>=2025.2.16.0",
    "uvicorn>=0.34.1",
]


[project.urls]
Repository = "https://github.com/soda92/bio-showcase"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
ignore-vcs = true
packages = ["bio_showcase", "bio_media", "bio_polls", "bio_primer"]
exclude = [
    "__pycache__",
    "*.pyc",
    "compile*.json",
    ".cache",
    "build",
    "*.syso",
    "*.exe~",
    "*.ico~",
    "*.pdb",
    "*.sqlite3",
]

[tool.hatch.build.targets.wheel]
ignore-vcs = true
packages = ["bio_showcase", "bio_media", "bio_polls", "bio_primer"]
# exclude is not needed because wheel is build from sdist in uv

# [tool.hatch.build.hooks.custom]

[project.scripts]
bio_run = "bio_showcase.main:main"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
ignore = ['F401']

[dependency-groups]
dev = ["jupyter>=1.1.1", "trayicons>=2025.5.12.1; sys_platform == 'win32'"]
